Add Statistics, Slim Docker Container + Fixed
Basic Statistics now exist. This should be improved upon
in future, however it works fine for now.
There is now also a simple UI for these under the /logs
endpoint on the client. This UI is just simple for now.

The Docker file now uses slim builds, which hopefully
decreases some relative sizes of the images.
It also now correctly builds the project, which may or may
not be the right way to do it, but it works.