feat: Add Filter List Parsing
This enables the use of Chumsky for parsing of filter lists. From
this the filters are loaded into the global Filter, where any
actions can be taken based on the rules of the filter.

The latter part of the above is still a WIP, as well as being able
to filter based on IP.